Addod a separate models package
